Change distance field on transport model

+++ transport/models.py
@@ -17,7 +17,7 @@ class Transport(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='request_owner')
     address = models.CharField(max_length=100)
     phone_number = models.CharField(max_length=15, null=True)
-    distance = models.IntegerField(null=True)
+    distance = models.DecimalField(max_digits=10,decimal_places=2,null=True)
     price = models.DecimalField(max_digits=10,decimal_places=2,null=True)
     is_paid = models.BooleanField(default=False)
     is_approved = models.BooleanField(default=False)
